requirements.txt
flask
pandas
scikit-learn
markdown
fpdf
google-generativeai